Fix(mesh): add static_cast<unsigned int> for indices.size() as it return size_t and not unsigned int

+++ src/mesh.cpp
@@ -71,6 +71,6 @@ void Mesh::draw(Shader &shader)
 
     // draw the mesh
     glBindVertexArray(VAO);
-    glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+    glDrawElements(GL_TRIANGLES, static_cast<unsigned int>(indices.size()), GL_UNSIGNED_INT, 0);
     glBindVertexArray(0);
 }
